Fix imports to avoid anti-pattern
