+++ packages/optimizely-sdk/lib/index.d.ts
@@ -17,6 +17,7 @@
 declare module '@optimizely/optimizely-sdk' {
   import { LogHandler, ErrorHandler } from '@optimizely/js-sdk-logging';
   import { Experiment, Variation } from '@optimizely/optimizely-sdk/lib/core/project_config';
+  import { UserProfileService, UserAttributes } from '@optimizely/optimizely-sdk/lib/core/decision_service';
   import * as enums from '@optimizely/optimizely-sdk/lib/utils/enums';
   import * as logging from '@optimizely/optimizely-sdk/lib/plugins/logger';
   export { enums, logging };
@@ -136,11 +137,6 @@ declare module '@optimizely/optimizely-sdk' {
     dispatchEvent: (event: Event, callback: () => void) => void;
   }
 
-  export interface UserProfileService {
-    lookup: (userId: string) => UserProfile;
-    save: (profile: UserProfile) => void;
-  }
-
   // NotificationCenter-related types
   export interface NotificationCenter {
     addNotificationListener<T extends ListenerPayload>(
@@ -165,12 +161,6 @@ declare module '@optimizely/optimizely-sdk' {
     logEvent: Event;
   }
 
-  export type UserAttributes = {
-    // TODO[OASIS-6649]: Don't use any type
-    // eslint-disable-next-line  @typescript-eslint/no-explicit-any
-    [name: string]: any;
-  };
-
   export type EventTags = {
     [key: string]: string | number | boolean;
   };
@@ -179,16 +169,6 @@ declare module '@optimizely/optimizely-sdk' {
     eventKey: string;
     eventTags: EventTags;
     logEvent: Event;
-  }
-
-  // Information about past bucketing decisions for a user.
-  export interface UserProfile {
-    user_id: string;
-    experiment_bucket_map: {
-      [experiment_id: string]: {
-        variation_id: string;
-      };
-    };
   }
 
   /**
